fix clause_head_functor/3 calculating incorrect arity in some cases
clauses matching the form (P, X => _) where X is a list are expanded
to prolog by make.pl with the variables in the list X being appended
to the head of the clause. The previous implementation of
clause_head_functor/3 did not take into account that the expansion
of this form changed the arity of the predicate.